(#57) Atualiza service de contain
